requirements.txt
gradio==4.44.0
google-generativeai==0.8.3
pillow==10.4.0
python-dotenv==1.0.1
streamlit==1.38.0
opencv-python==4.10.0.84
numpy==1.26.4
